pyproject.toml
[project]
name = "lectures"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "en-core-web-sm",
    "fastparquet>=2024.11.0",
    "groq>=0.33.0",
    "jupyter>=1.1.1",
    "lxml>=5.4.0",
    "matplotlib>=3.10.3",
    "ollama>=0.6.0",
    "openai>=2.6.1",
    "pandas>=2.2.3",
    "plotly>=6.3.1",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "spacy>=3.8.7",
    "statsmodels>=0.14.4",
    "tiktoken>=0.12.0",
    "torch>=2.9.0",
    "transformers>=4.57.1",
    "umap>=0.1.1",
    "wordcloud>=1.9.4",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0.tar.gz" }
